IDEV-468 Fix default value for numberOfPhases, serialization of 0 value and OCPP 1.6 limit as Int

+++ ocpp-1-6-api-adapter/src/main/kotlin/com/izivia/ocpp/adapter16/mapper/GetCompositeScheduleMapper.kt
@@ -37,7 +37,7 @@ abstract class GetCompositeScheduleMapper {
                 chargingSchedulePeriod = schedule.chargingSchedulePeriod.map {
                     ChargingSchedulePeriod(
                         it.startPeriod,
-                        it.limit,
+                        it.limit.toInt(),
                         it.numberPhases
                     )
                 }
